Move Log events to the new format
